LABA 7

+++ MUHITAV_Autoservice/SERVESYS.cs
@@ -25,8 +25,8 @@ namespace MUHITAV_Autoservice
         public string Title { get; set; }
         public string MainImagePath { get; set; }
         public string DurationInSeconds { get; set; }
-        public decimal Cost { get; set; }
-        public Nullable<int> Discount { get; set; }
+        public short Cost { get; set; }
+        public Nullable<byte> Discount { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<ClientService> ClientService { get; set; }
